Refactor testing framework tests by component
