+++ pytest.ini
@@ -1,2 +1,2 @@
 [pytest]
-testpaths = tests/unit
+testpaths = tests
