+++ tests/test_testing_framework_fixtures.py
@@ -1,0 +1,325 @@
+#!/usr/bin/env python3
+"""
+Testing Framework Fixture Tests - TDD Implementation
+====================================================
+
+Tests for fixture-based integration test classes and suites.
+"""
+
+import pytest
+
+from src.core.testing_framework import (
+    TestStatus,
+    TestType,
+    TestPriority,
+    TestResult,
+    BaseIntegrationTest,
+    CrossSystemCommunicationTest,
+    ServiceIntegrationTest,
+    DatabaseIntegrationTest,
+    IntegrationTestSuite,
+    TestEnvironment,
+)
+
+
+@pytest.fixture
+def communication_test():
+    """Create a test communication test instance."""
+    return CrossSystemCommunicationTest("test_communication", TestPriority.HIGH)
+
+
+@pytest.fixture
+def service_test():
+    """Create a test service integration test instance."""
+    return ServiceIntegrationTest("test_service", TestPriority.NORMAL)
+
+
+@pytest.fixture
+def database_test():
+    """Create a test database integration test instance."""
+    return DatabaseIntegrationTest("test_database", TestPriority.NORMAL)
+
+
+@pytest.fixture
+def test_suite():
+    """Create a test suite instance."""
+    return IntegrationTestSuite("Test Suite", "Test description")
+
+
+@pytest.fixture
+def sample_test():
+    """Create a sample test instance."""
+    return CrossSystemCommunicationTest("Sample Test", TestPriority.NORMAL)
+
+
+class TestBaseIntegrationTest:
+    """Test BaseIntegrationTest abstract class following TDD principles"""
+
+    @pytest.mark.unit
+    def test_abstract_class_instantiation(self):
+        """Test that BaseIntegrationTest cannot be instantiated."""
+        with pytest.raises(TypeError):
+            BaseIntegrationTest("test_name", TestType.INTEGRATION)
+
+    @pytest.mark.unit
+    def test_abstract_methods(self):
+        """Test that abstract methods are defined."""
+        # This test ensures the abstract methods exist
+        assert hasattr(BaseIntegrationTest, "setup")
+        assert hasattr(BaseIntegrationTest, "execute")
+        assert hasattr(BaseIntegrationTest, "cleanup")
+        assert hasattr(BaseIntegrationTest, "validate")
+
+
+class TestCrossSystemCommunicationTest:
+    """Test CrossSystemCommunicationTest following TDD principles"""
+
+    @pytest.mark.unit
+    def test_test_initialization(self, communication_test):
+        """Test test initialization."""
+        assert communication_test.test_name == "test_communication"
+        assert communication_test.test_type == TestType.INTEGRATION
+        assert communication_test.priority == TestPriority.HIGH
+        assert communication_test.status == TestStatus.PENDING
+        assert len(communication_test.communication_systems) == 0
+
+    @pytest.mark.asyncio
+    async def test_setup_phase(self, communication_test):
+        """Test setup phase execution."""
+        result = await communication_test.setup()
+        
+        assert result is True
+        assert len(communication_test.communication_systems) == 3
+        assert len(communication_test.connection_status) == 3
+        assert "system_a" in communication_test.connection_status
+        assert communication_test.connection_status["system_a"] == "connected"
+
+    @pytest.mark.asyncio
+    async def test_execute_phase(self, communication_test):
+        """Test execute phase execution."""
+        # Setup first
+        await communication_test.setup()
+        
+        # Execute
+        result = await communication_test.execute()
+        
+        assert result is True
+        assert communication_test.assertions_passed > 0
+        assert communication_test.assertions_failed >= 0
+
+    @pytest.mark.asyncio
+    async def test_cleanup_phase(self, communication_test):
+        """Test cleanup phase execution."""
+        # Setup and execute first
+        await communication_test.setup()
+        await communication_test.execute()
+        
+        # Cleanup
+        result = await communication_test.cleanup()
+        
+        assert result is True
+        assert len(communication_test.communication_systems) == 0
+        assert len(communication_test.connection_status) == 0
+        assert len(communication_test.message_queue) == 0
+
+    @pytest.mark.asyncio
+    async def test_validate_phase(self, communication_test):
+        """Test validate phase execution."""
+        # Setup and execute first
+        await communication_test.setup()
+        await communication_test.execute()
+        
+        # Validate
+        result = await communication_test.validate()
+        
+        assert result is True
+        assert "success_rate" in communication_test.metrics
+        assert communication_test.metrics["success_rate"] >= 0.8
+
+    @pytest.mark.asyncio
+    async def test_complete_test_lifecycle(self, communication_test):
+        """Test complete test lifecycle execution."""
+        result = await communication_test.run()
+        
+        assert isinstance(result, TestResult)
+        assert result.test_name == "test_communication"
+        assert result.test_type == TestType.INTEGRATION
+        assert result.status in [TestStatus.PASSED, TestStatus.FAILED]
+        assert result.duration > 0
+
+
+class TestServiceIntegrationTest:
+    """Test ServiceIntegrationTest following TDD principles"""
+
+    @pytest.mark.unit
+    def test_test_initialization(self, service_test):
+        """Test test initialization."""
+        assert service_test.test_name == "test_service"
+        assert service_test.test_type == TestType.INTEGRATION
+        assert service_test.priority == TestPriority.NORMAL
+        assert len(service_test.services) == 0
+
+    @pytest.mark.asyncio
+    async def test_setup_phase(self, service_test):
+        """Test setup phase execution."""
+        result = await service_test.setup()
+        
+        assert result is True
+        assert len(service_test.services) == 3
+        assert len(service_test.service_status) == 3
+        assert "auth_service" in service_test.service_status
+        assert service_test.service_status["auth_service"] == "healthy"
+
+    @pytest.mark.asyncio
+    async def test_execute_phase(self, service_test):
+        """Test execute phase execution."""
+        # Setup first
+        await service_test.setup()
+        
+        # Execute
+        result = await service_test.execute()
+        
+        assert result is True
+        assert service_test.assertions_passed > 0
+        assert len(service_test.api_calls) > 0
+
+    @pytest.mark.asyncio
+    async def test_complete_test_lifecycle(self, service_test):
+        """Test complete test lifecycle execution."""
+        result = await service_test.run()
+        
+        assert isinstance(result, TestResult)
+        assert result.test_name == "test_service"
+        assert result.test_type == TestType.INTEGRATION
+        assert result.status in [TestStatus.PASSED, TestStatus.FAILED]
+        assert result.duration > 0
+
+
+class TestDatabaseIntegrationTest:
+    """Test DatabaseIntegrationTest following TDD principles"""
+
+    @pytest.mark.unit
+    def test_test_initialization(self, database_test):
+        """Test test initialization."""
+        assert database_test.test_name == "test_database"
+        assert database_test.test_type == TestType.INTEGRATION
+        assert database_test.priority == TestPriority.NORMAL
+        assert len(database_test.database_connections) == 0
+
+    @pytest.mark.asyncio
+    async def test_setup_phase(self, database_test):
+        """Test setup phase execution."""
+        result = await database_test.setup()
+        
+        assert result is True
+        assert len(database_test.database_connections) == 3
+        assert len(database_test.test_queries) == 3
+        assert "main_db" in [db["name"] for db in database_test.database_connections]
+
+    @pytest.mark.asyncio
+    async def test_execute_phase(self, database_test):
+        """Test execute phase execution."""
+        # Setup first
+        await database_test.setup()
+        
+        # Execute
+        result = await database_test.execute()
+        
+        assert result is True
+        assert database_test.assertions_passed > 0
+        assert len(database_test.query_results) > 0
+
+    @pytest.mark.asyncio
+    async def test_complete_test_lifecycle(self, database_test):
+        """Test complete test lifecycle execution."""
+        result = await database_test.run()
+        
+        assert isinstance(result, TestResult)
+        assert result.test_name == "test_database"
+        assert result.test_type == TestType.INTEGRATION
+        assert result.status in [TestStatus.PASSED, TestStatus.FAILED]
+        assert result.duration > 0
+
+
+class TestIntegrationTestSuite:
+    """Test IntegrationTestSuite following TDD principles"""
+
+    @pytest.mark.unit
+    def test_suite_initialization(self, test_suite):
+        """Test suite initialization."""
+        assert test_suite.name == "Test Suite"
+        assert test_suite.description == "Test description"
+        assert len(test_suite.tests) == 0
+        assert test_suite.parallel_execution is False
+        assert test_suite.max_parallel_tests == 5
+
+    @pytest.mark.unit
+    def test_add_test(self, test_suite, sample_test):
+        """Test adding a test to the suite."""
+        test_suite.add_test(sample_test)
+        
+        assert len(test_suite.tests) == 1
+        assert test_suite.tests[0] == sample_test
+
+    @pytest.mark.unit
+    def test_add_invalid_test(self, test_suite):
+        """Test adding an invalid test to the suite."""
+        with pytest.raises(ValueError, match="Test must be an instance of BaseIntegrationTest"):
+            test_suite.add_test("invalid_test")
+
+    @pytest.mark.unit
+    def test_remove_test(self, test_suite, sample_test):
+        """Test removing a test from the suite."""
+        test_suite.add_test(sample_test)
+        assert len(test_suite.tests) == 1
+        
+        result = test_suite.remove_test("Sample Test")
+        assert result is True
+        assert len(test_suite.tests) == 0
+
+    @pytest.mark.unit
+    def test_remove_nonexistent_test(self, test_suite):
+        """Test removing a nonexistent test from the suite."""
+        result = test_suite.remove_test("Nonexistent Test")
+        assert result is False
+
+    @pytest.mark.unit
+    def test_set_environment(self, test_suite):
+        """Test setting environment for the suite."""
+        environment = TestEnvironment("env_123", "Test Env", "Test environment")
+        environment.parallel_execution = True
+        environment.max_parallel_tests = 10
+        
+        test_suite.set_environment(environment)
+        
+        assert test_suite.parallel_execution is True
+        assert test_suite.max_parallel_tests == 10
+
+    @pytest.mark.asyncio
+    async def test_run_empty_suite(self, test_suite):
+        """Test running an empty suite."""
+        results = await test_suite.run_suite()
+        
+        assert results == []
+
+    @pytest.mark.asyncio
+    async def test_run_suite_with_tests(self, test_suite, sample_test):
+        """Test running a suite with tests."""
+        test_suite.add_test(sample_test)
+        
+        results = await test_suite.run_suite()
+        
+        assert len(results) == 1
+        assert isinstance(results[0], TestResult)
+
+    @pytest.mark.unit
+    def test_get_summary_empty_suite(self, test_suite):
+        """Test getting summary for empty suite."""
+        summary = test_suite.get_summary()
+        
+        assert summary["name"] == "Test Suite"
+        assert summary["total_tests"] == 0
+        assert summary["passed"] == 0
+        assert summary["success_rate"] == 0.0
+
+
