--- tests/test_testing_framework_modular.py
@@ -1,974 +1,0 @@
-#!/usr/bin/env python3
-"""
-🧪 Testing Framework Modular Tests - TDD Implementation
-=======================================================
-
-Test-Driven Development implementation for modular testing framework system.
-Tests the refactored V2-compliant testing framework components following OOP and SRP principles.
-
-Author: Agent-2 (Architecture & Design Specialist)
-License: MIT
-"""
-
-import pytest
-import asyncio
-import tempfile
-
-from src.utils.stability_improvements import stability_manager, safe_import
-from pathlib import Path
-from unittest.mock import Mock, patch, AsyncMock
-
-# Import our new modular components
-from src.core.testing_framework import (
-    TestStatus, TestType, TestPriority,
-    TestResult, TestScenario, TestEnvironment
-)
-from src.core.testing_framework import (
-    BaseIntegrationTest, CrossSystemCommunicationTest,
-    ServiceIntegrationTest, DatabaseIntegrationTest
-)
-from src.core.testing_framework import (
-    IntegrationTestSuite, IntegrationTestRunner,
-    TestExecutor, TestOrchestrator
-)
-from src.core.testing_framework import TestingFrameworkCLI
-
-
-class TestTestStatus:
-    """Test TestStatus enum following TDD principles"""
-
-    @pytest.mark.unit
-    def test_status_values(self):
-        """Test that all status values are valid."""
-        statuses = [
-            TestStatus.PENDING,
-            TestStatus.RUNNING,
-            TestStatus.PASSED,
-            TestStatus.FAILED,
-            TestStatus.SKIPPED,
-            TestStatus.TIMEOUT,
-            TestStatus.ERROR,
-        ]
-
-        for status in statuses:
-            assert status.value in [
-                "pending",
-                "running",
-                "passed",
-                "failed",
-                "skipped",
-                "timeout",
-                "error",
-            ]
-
-
-class TestTestType:
-    """Test TestType enum following TDD principles"""
-
-    @pytest.mark.unit
-    def test_type_values(self):
-        """Test that all type values are valid."""
-        types = [
-            TestType.UNIT,
-            TestType.INTEGRATION,
-            TestType.END_TO_END,
-            TestType.PERFORMANCE,
-            TestType.LOAD,
-            TestType.STRESS,
-            TestType.SECURITY,
-            TestType.COMPATIBILITY,
-        ]
-
-        for test_type in types:
-            assert test_type.value in [
-                "unit",
-                "integration",
-                "end_to_end",
-                "performance",
-                "load",
-                "stress",
-                "security",
-                "compatibility",
-            ]
-
-
-class TestTestPriority:
-    """Test TestPriority enum following TDD principles"""
-
-    @pytest.mark.unit
-    def test_priority_values(self):
-        """Test that all priority values are valid."""
-        priorities = [
-            TestPriority.LOW,
-            TestPriority.NORMAL,
-            TestPriority.HIGH,
-            TestPriority.CRITICAL,
-        ]
-
-        for priority in priorities:
-            assert priority.value in [1, 2, 3, 4]
-
-    @pytest.mark.unit
-    def test_priority_ordering(self):
-        """Test that priority values are properly ordered."""
-        assert TestPriority.LOW.value < TestPriority.NORMAL.value
-        assert TestPriority.NORMAL.value < TestPriority.HIGH.value
-        assert TestPriority.HIGH.value < TestPriority.CRITICAL.value
-
-
-class TestTestResult:
-    """Test TestResult dataclass following TDD principles"""
-
-    @pytest.mark.unit
-    def test_result_creation_with_all_parameters(self):
-        """Test creating TestResult with all parameters."""
-        result = TestResult(
-            test_id="test_123",
-            test_name="Test Name",
-            test_type=TestType.INTEGRATION,
-            status=TestStatus.PASSED,
-            start_time=100.0,
-            end_time=110.0,
-            duration=10.0,
-        )
-
-        assert result.test_id == "test_123"
-        assert result.test_name == "Test Name"
-        assert result.test_type == TestType.INTEGRATION
-        assert result.status == TestStatus.PASSED
-        assert result.start_time == 100.0
-        assert result.end_time == 110.0
-        assert result.duration == 10.0
-        assert result.error_message is None
-        assert result.error_traceback is None
-        assert result.metrics == {}
-        assert result.logs == []
-        assert result.assertions_passed == 0
-        assert result.assertions_failed == 0
-        assert result.test_data == {}
-
-    @pytest.mark.unit
-    def test_result_with_optional_fields(self):
-        """Test creating TestResult with optional fields."""
-        result = TestResult(
-            test_id="test_456",
-            test_name="Test With Data",
-            test_type=TestType.PERFORMANCE,
-            status=TestStatus.FAILED,
-            start_time=200.0,
-            end_time=205.0,
-            duration=5.0,
-            error_message="Test failed due to timeout",
-            error_traceback="Traceback...",
-            metrics={"response_time": 2.5, "throughput": 100},
-            logs=["Starting test", "Test completed"],
-            assertions_passed=5,
-            assertions_failed=1,
-            test_data={"input": "test_input", "output": "test_output"},
-        )
-
-        assert result.error_message == "Test failed due to timeout"
-        assert result.error_traceback == "Traceback..."
-        assert result.metrics["response_time"] == 2.5
-        assert result.metrics["throughput"] == 100
-        assert len(result.logs) == 2
-        assert result.assertions_passed == 5
-        assert result.assertions_failed == 1
-        assert result.test_data["input"] == "test_input"
-
-    @pytest.mark.unit
-    def test_result_auto_duration_calculation(self):
-        """Test that duration is auto-calculated if not provided."""
-        result = TestResult(
-            test_id="test_789",
-            test_name="Auto Duration Test",
-            test_type=TestType.UNIT,
-            status=TestStatus.PASSED,
-            start_time=100.0,
-            end_time=110.0,
-            duration=0.0,  # Will be auto-calculated
-        )
-
-        assert result.duration == 10.0  # Auto-calculated
-
-
-class TestTestScenario:
-    """Test TestScenario dataclass following TDD principles"""
-
-    @pytest.mark.unit
-    def test_scenario_creation_with_all_parameters(self):
-        """Test creating TestScenario with all parameters."""
-        scenario = TestScenario(
-            scenario_id="scenario_123",
-            name="Test Scenario",
-            description="A test scenario description",
-            test_type=TestType.INTEGRATION,
-            priority=TestPriority.HIGH,
-        )
-
-        assert scenario.scenario_id == "scenario_123"
-        assert scenario.name == "Test Scenario"
-        assert scenario.description == "A test scenario description"
-        assert scenario.test_type == TestType.INTEGRATION
-        assert scenario.priority == TestPriority.HIGH
-        assert scenario.timeout == 300.0
-        assert scenario.retry_count == 0
-        assert scenario.max_retries == 3
-        assert scenario.dependencies == []
-        assert scenario.setup_steps == []
-        assert scenario.test_steps == []
-        assert scenario.cleanup_steps == []
-        assert scenario.assertions == []
-        assert scenario.metadata == {}
-
-    @pytest.mark.unit
-    def test_scenario_with_custom_values(self):
-        """Test creating a scenario with custom values."""
-        scenario = TestScenario(
-            scenario_id="custom_scenario",
-            name="Custom Scenario",
-            description="Custom scenario with steps",
-            test_type=TestType.END_TO_END,
-            priority=TestPriority.CRITICAL,
-            timeout=600.0,
-            retry_count=1,
-            max_retries=5,
-            dependencies=["dependency1", "dependency2"],
-            metadata={"version": "1.0.0", "environment": "production"},
-        )
-
-        assert scenario.timeout == 600.0
-        assert scenario.retry_count == 1
-        assert scenario.max_retries == 5
-        assert len(scenario.dependencies) == 2
-        assert scenario.metadata["version"] == "1.0.0"
-
-    @pytest.mark.unit
-    def test_scenario_validation_timeout(self):
-        """Test scenario validation for invalid timeout."""
-        with pytest.raises(ValueError, match="Timeout must be positive"):
-            TestScenario(
-                scenario_id="invalid",
-                name="Invalid Scenario",
-                description="Invalid scenario",
-                test_type=TestType.INTEGRATION,
-                priority=TestPriority.NORMAL,
-                timeout=0.0
-            )
-
-    @pytest.mark.unit
-    def test_scenario_validation_retries(self):
-        """Test scenario validation for invalid retry count."""
-        with pytest.raises(ValueError, match="Retry count cannot exceed max retries"):
-            TestScenario(
-                scenario_id="invalid",
-                name="Invalid Scenario",
-                description="Invalid scenario",
-                test_type=TestType.INTEGRATION,
-                priority=TestPriority.NORMAL,
-                retry_count=5,
-                max_retries=3
-            )
-
-
-class TestTestEnvironment:
-    """Test TestEnvironment dataclass following TDD principles"""
-
-    @pytest.mark.unit
-    def test_environment_creation_with_all_parameters(self):
-        """Test creating TestEnvironment with all parameters."""
-        environment = TestEnvironment(
-            environment_id="env_123",
-            name="Test Environment",
-            description="A test environment",
-        )
-
-        assert environment.environment_id == "env_123"
-        assert environment.name == "Test Environment"
-        assert environment.description == "A test environment"
-        assert environment.systems == []
-        assert environment.services == []
-        assert environment.config_overrides == {}
-        assert environment.cleanup_on_exit is True
-        assert environment.parallel_execution is False
-        assert environment.max_parallel_tests == 5
-
-    @pytest.mark.unit
-    def test_environment_with_custom_values(self):
-        """Test creating an environment with custom values."""
-        environment = TestEnvironment(
-            environment_id="custom_env",
-            name="Custom Environment",
-            description="Custom environment with settings",
-            cleanup_on_exit=False,
-            parallel_execution=True,
-            max_parallel_tests=10,
-        )
-
-        assert environment.cleanup_on_exit is False
-        assert environment.parallel_execution is True
-        assert environment.max_parallel_tests == 10
-
-    @pytest.mark.unit
-    def test_environment_validation_parallel_tests(self):
-        """Test environment validation for invalid parallel tests."""
-        with pytest.raises(ValueError, match="Max parallel tests must be positive"):
-            TestEnvironment(
-                environment_id="invalid",
-                name="Invalid Environment",
-                description="Invalid environment",
-                max_parallel_tests=0
-            )
-
-    @pytest.mark.unit
-    def test_environment_auto_disable_parallel(self):
-        """Test that parallel execution is auto-disabled for single test."""
-        environment = TestEnvironment(
-            environment_id="auto_disable",
-            name="Auto Disable Environment",
-            description="Environment that auto-disables parallel execution",
-            parallel_execution=True,
-            max_parallel_tests=1
-        )
-
-        assert environment.parallel_execution is False
-
-
-class TestBaseIntegrationTest:
-    """Test BaseIntegrationTest abstract class following TDD principles"""
-
-    @pytest.mark.unit
-    def test_abstract_class_instantiation(self):
-        """Test that BaseIntegrationTest cannot be instantiated."""
-        with pytest.raises(TypeError):
-            BaseIntegrationTest("test_name", TestType.INTEGRATION)
-
-    @pytest.mark.unit
-    def test_abstract_methods(self):
-        """Test that abstract methods are defined."""
-        # This test ensures the abstract methods exist
-        assert hasattr(BaseIntegrationTest, "setup")
-        assert hasattr(BaseIntegrationTest, "execute")
-        assert hasattr(BaseIntegrationTest, "cleanup")
-        assert hasattr(BaseIntegrationTest, "validate")
-
-
-class TestCrossSystemCommunicationTest:
-    """Test CrossSystemCommunicationTest following TDD principles"""
-
-    @pytest.fixture
-    def communication_test(self):
-        """Create a test communication test instance."""
-        return CrossSystemCommunicationTest("test_communication", TestPriority.HIGH)
-
-    @pytest.mark.unit
-    def test_test_initialization(self, communication_test):
-        """Test test initialization."""
-        assert communication_test.test_name == "test_communication"
-        assert communication_test.test_type == TestType.INTEGRATION
-        assert communication_test.priority == TestPriority.HIGH
-        assert communication_test.status == TestStatus.PENDING
-        assert len(communication_test.communication_systems) == 0
-
-    @pytest.mark.asyncio
-    async def test_setup_phase(self, communication_test):
-        """Test setup phase execution."""
-        result = await communication_test.setup()
-        
-        assert result is True
-        assert len(communication_test.communication_systems) == 3
-        assert len(communication_test.connection_status) == 3
-        assert "system_a" in communication_test.connection_status
-        assert communication_test.connection_status["system_a"] == "connected"
-
-    @pytest.mark.asyncio
-    async def test_execute_phase(self, communication_test):
-        """Test execute phase execution."""
-        # Setup first
-        await communication_test.setup()
-        
-        # Execute
-        result = await communication_test.execute()
-        
-        assert result is True
-        assert communication_test.assertions_passed > 0
-        assert communication_test.assertions_failed >= 0
-
-    @pytest.mark.asyncio
-    async def test_cleanup_phase(self, communication_test):
-        """Test cleanup phase execution."""
-        # Setup and execute first
-        await communication_test.setup()
-        await communication_test.execute()
-        
-        # Cleanup
-        result = await communication_test.cleanup()
-        
-        assert result is True
-        assert len(communication_test.communication_systems) == 0
-        assert len(communication_test.connection_status) == 0
-        assert len(communication_test.message_queue) == 0
-
-    @pytest.mark.asyncio
-    async def test_validate_phase(self, communication_test):
-        """Test validate phase execution."""
-        # Setup and execute first
-        await communication_test.setup()
-        await communication_test.execute()
-        
-        # Validate
-        result = await communication_test.validate()
-        
-        assert result is True
-        assert "success_rate" in communication_test.metrics
-        assert communication_test.metrics["success_rate"] >= 0.8
-
-    @pytest.mark.asyncio
-    async def test_complete_test_lifecycle(self, communication_test):
-        """Test complete test lifecycle execution."""
-        result = await communication_test.run()
-        
-        assert isinstance(result, TestResult)
-        assert result.test_name == "test_communication"
-        assert result.test_type == TestType.INTEGRATION
-        assert result.status in [TestStatus.PASSED, TestStatus.FAILED]
-        assert result.duration > 0
-
-
-class TestServiceIntegrationTest:
-    """Test ServiceIntegrationTest following TDD principles"""
-
-    @pytest.fixture
-    def service_test(self):
-        """Create a test service integration test instance."""
-        return ServiceIntegrationTest("test_service", TestPriority.NORMAL)
-
-    @pytest.mark.unit
-    def test_test_initialization(self, service_test):
-        """Test test initialization."""
-        assert service_test.test_name == "test_service"
-        assert service_test.test_type == TestType.INTEGRATION
-        assert service_test.priority == TestPriority.NORMAL
-        assert len(service_test.services) == 0
-
-    @pytest.mark.asyncio
-    async def test_setup_phase(self, service_test):
-        """Test setup phase execution."""
-        result = await service_test.setup()
-        
-        assert result is True
-        assert len(service_test.services) == 3
-        assert len(service_test.service_status) == 3
-        assert "auth_service" in service_test.service_status
-        assert service_test.service_status["auth_service"] == "healthy"
-
-    @pytest.mark.asyncio
-    async def test_execute_phase(self, service_test):
-        """Test execute phase execution."""
-        # Setup first
-        await service_test.setup()
-        
-        # Execute
-        result = await service_test.execute()
-        
-        assert result is True
-        assert service_test.assertions_passed > 0
-        assert len(service_test.api_calls) > 0
-
-    @pytest.mark.asyncio
-    async def test_complete_test_lifecycle(self, service_test):
-        """Test complete test lifecycle execution."""
-        result = await service_test.run()
-        
-        assert isinstance(result, TestResult)
-        assert result.test_name == "test_service"
-        assert result.test_type == TestType.INTEGRATION
-        assert result.status in [TestStatus.PASSED, TestStatus.FAILED]
-        assert result.duration > 0
-
-
-class TestDatabaseIntegrationTest:
-    """Test DatabaseIntegrationTest following TDD principles"""
-
-    @pytest.fixture
-    def database_test(self):
-        """Create a test database integration test instance."""
-        return DatabaseIntegrationTest("test_database", TestPriority.NORMAL)
-
-    @pytest.mark.unit
-    def test_test_initialization(self, database_test):
-        """Test test initialization."""
-        assert database_test.test_name == "test_database"
-        assert database_test.test_type == TestType.INTEGRATION
-        assert database_test.priority == TestPriority.NORMAL
-        assert len(database_test.database_connections) == 0
-
-    @pytest.mark.asyncio
-    async def test_setup_phase(self, database_test):
-        """Test setup phase execution."""
-        result = await database_test.setup()
-        
-        assert result is True
-        assert len(database_test.database_connections) == 3
-        assert len(database_test.test_queries) == 3
-        assert "main_db" in [db["name"] for db in database_test.database_connections]
-
-    @pytest.mark.asyncio
-    async def test_execute_phase(self, database_test):
-        """Test execute phase execution."""
-        # Setup first
-        await database_test.setup()
-        
-        # Execute
-        result = await database_test.execute()
-        
-        assert result is True
-        assert database_test.assertions_passed > 0
-        assert len(database_test.query_results) > 0
-
-    @pytest.mark.asyncio
-    async def test_complete_test_lifecycle(self, database_test):
-        """Test complete test lifecycle execution."""
-        result = await database_test.run()
-        
-        assert isinstance(result, TestResult)
-        assert result.test_name == "test_database"
-        assert result.test_type == TestType.INTEGRATION
-        assert result.status in [TestStatus.PASSED, TestStatus.FAILED]
-        assert result.duration > 0
-
-
-class TestIntegrationTestSuite:
-    """Test IntegrationTestSuite following TDD principles"""
-
-    @pytest.fixture
-    def test_suite(self):
-        """Create a test suite instance."""
-        return IntegrationTestSuite("Test Suite", "Test description")
-
-    @pytest.fixture
-    def sample_test(self):
-        """Create a sample test instance."""
-        return CrossSystemCommunicationTest("Sample Test", TestPriority.NORMAL)
-
-    @pytest.mark.unit
-    def test_suite_initialization(self, test_suite):
-        """Test suite initialization."""
-        assert test_suite.name == "Test Suite"
-        assert test_suite.description == "Test description"
-        assert len(test_suite.tests) == 0
-        assert test_suite.parallel_execution is False
-        assert test_suite.max_parallel_tests == 5
-
-    @pytest.mark.unit
-    def test_add_test(self, test_suite, sample_test):
-        """Test adding a test to the suite."""
-        test_suite.add_test(sample_test)
-        
-        assert len(test_suite.tests) == 1
-        assert test_suite.tests[0] == sample_test
-
-    @pytest.mark.unit
-    def test_add_invalid_test(self, test_suite):
-        """Test adding an invalid test to the suite."""
-        with pytest.raises(ValueError, match="Test must be an instance of BaseIntegrationTest"):
-            test_suite.add_test("invalid_test")
-
-    @pytest.mark.unit
-    def test_remove_test(self, test_suite, sample_test):
-        """Test removing a test from the suite."""
-        test_suite.add_test(sample_test)
-        assert len(test_suite.tests) == 1
-        
-        result = test_suite.remove_test("Sample Test")
-        assert result is True
-        assert len(test_suite.tests) == 0
-
-    @pytest.mark.unit
-    def test_remove_nonexistent_test(self, test_suite):
-        """Test removing a nonexistent test from the suite."""
-        result = test_suite.remove_test("Nonexistent Test")
-        assert result is False
-
-    @pytest.mark.unit
-    def test_set_environment(self, test_suite):
-        """Test setting environment for the suite."""
-        environment = TestEnvironment("env_123", "Test Env", "Test environment")
-        environment.parallel_execution = True
-        environment.max_parallel_tests = 10
-        
-        test_suite.set_environment(environment)
-        
-        assert test_suite.parallel_execution is True
-        assert test_suite.max_parallel_tests == 10
-
-    @pytest.mark.asyncio
-    async def test_run_empty_suite(self, test_suite):
-        """Test running an empty suite."""
-        results = await test_suite.run_suite()
-        
-        assert results == []
-
-    @pytest.mark.asyncio
-    async def test_run_suite_with_tests(self, test_suite, sample_test):
-        """Test running a suite with tests."""
-        test_suite.add_test(sample_test)
-        
-        results = await test_suite.run_suite()
-        
-        assert len(results) == 1
-        assert isinstance(results[0], TestResult)
-
-    @pytest.mark.unit
-    def test_get_summary_empty_suite(self, test_suite):
-        """Test getting summary for empty suite."""
-        summary = test_suite.get_summary()
-        
-        assert summary["name"] == "Test Suite"
-        assert summary["total_tests"] == 0
-        assert summary["passed"] == 0
-        assert summary["success_rate"] == 0.0
-
-
-class TestIntegrationTestRunner:
-    """Test IntegrationTestRunner following TDD principles"""
-
-    @pytest.fixture
-    def test_runner(self):
-        """Create a test runner instance."""
-        return IntegrationTestRunner()
-
-    @pytest.fixture
-    def sample_suite(self):
-        """Create a sample test suite."""
-        suite = IntegrationTestSuite("Sample Suite", "Sample description")
-        suite.add_test(CrossSystemCommunicationTest("Test 1", TestPriority.NORMAL))
-        return suite
-
-    @pytest.mark.unit
-    def test_runner_initialization(self, test_runner):
-        """Test runner initialization."""
-        assert len(test_runner.test_suites) == 0
-        assert test_runner.global_config["parallel_suites"] is False
-        assert test_runner.global_config["max_parallel_suites"] == 3
-
-    @pytest.mark.unit
-    def test_add_test_suite(self, test_runner, sample_suite):
-        """Test adding a test suite to the runner."""
-        test_runner.add_test_suite(sample_suite)
-        
-        assert len(test_runner.test_suites) == 1
-        assert test_runner.test_suites[0] == sample_suite
-
-    @pytest.mark.unit
-    def test_add_invalid_suite(self, test_runner):
-        """Test adding an invalid suite to the runner."""
-        with pytest.raises(ValueError, match="Suite must be an instance of IntegrationTestSuite"):
-            test_runner.add_test_suite("invalid_suite")
-
-    @pytest.mark.unit
-    def test_remove_test_suite(self, test_runner, sample_suite):
-        """Test removing a test suite from the runner."""
-        test_runner.add_test_suite(sample_suite)
-        assert len(test_runner.test_suites) == 1
-        
-        result = test_runner.remove_test_suite("Sample Suite")
-        assert result is True
-        assert len(test_runner.test_suites) == 0
-
-    @pytest.mark.unit
-    def test_remove_nonexistent_suite(self, test_runner):
-        """Test removing a nonexistent suite from the runner."""
-        result = test_runner.remove_test_suite("Nonexistent Suite")
-        assert result is False
-
-    @pytest.mark.unit
-    def test_set_global_config(self, test_runner):
-        """Test setting global configuration."""
-        config = {"parallel_suites": True, "max_parallel_suites": 5}
-        test_runner.set_global_config(config)
-        
-        assert test_runner.global_config["parallel_suites"] is True
-        assert test_runner.global_config["max_parallel_suites"] == 5
-
-    @pytest.mark.asyncio
-    async def test_run_empty_runner(self, test_runner):
-        """Test running an empty runner."""
-        results = await test_runner.run_all_suites()
-        
-        assert results == {}
-
-    @pytest.mark.asyncio
-    async def test_run_runner_with_suites(self, test_runner, sample_suite):
-        """Test running a runner with suites."""
-        test_runner.add_test_suite(sample_suite)
-        
-        results = await test_runner.run_all_suites()
-        
-        assert "Sample Suite" in results
-        assert len(results["Sample Suite"]) > 0
-
-    @pytest.mark.unit
-    def test_get_global_summary_empty_runner(self, test_runner):
-        """Test getting global summary for empty runner."""
-        summary = test_runner.get_global_summary()
-        
-        assert summary["total_suites"] == 0
-        assert summary["total_tests"] == 0
-        assert summary["overall_success_rate"] == 0.0
-
-
-class TestTestExecutor:
-    """Test TestExecutor following TDD principles"""
-
-    @pytest.fixture
-    def test_executor(self):
-        """Create a test executor instance."""
-        return TestExecutor()
-
-    @pytest.fixture
-    def sample_test(self):
-        """Create a sample test instance."""
-        return CrossSystemCommunicationTest("Executor Test", TestPriority.NORMAL)
-
-    @pytest.mark.unit
-    def test_executor_initialization(self, test_executor):
-        """Test executor initialization."""
-        assert len(test_executor.execution_history) == 0
-
-    @pytest.mark.asyncio
-    async def test_execute_test_success(self, test_executor, sample_test):
-        """Test successful test execution."""
-        result = await test_executor.execute_test(sample_test, timeout=60.0)
-        
-        assert isinstance(result, TestResult)
-        assert result.test_name == "Executor Test"
-        assert result.status in [TestStatus.PASSED, TestStatus.FAILED]
-        assert len(test_executor.execution_history) == 1
-
-    @pytest.mark.asyncio
-    async def test_execute_test_timeout(self, test_executor, sample_test):
-        """Test test execution with timeout."""
-        # Mock the test to take longer than timeout
-        with patch.object(sample_test, 'run', side_effect=asyncio.sleep(2)):
-            result = await test_executor.execute_test(sample_test, timeout=0.1)
-        
-        assert result.status == TestStatus.TIMEOUT
-        assert "timed out" in result.error_message
-
-    @pytest.mark.unit
-    def test_get_execution_history(self, test_executor):
-        """Test getting execution history."""
-        history = test_executor.get_execution_history()
-        
-        assert isinstance(history, list)
-        assert len(history) == 0
-
-    @pytest.mark.unit
-    def test_clear_history(self, test_executor):
-        """Test clearing execution history."""
-        # Add some mock history
-        test_executor.execution_history = [Mock(), Mock()]
-        assert len(test_executor.execution_history) == 2
-        
-        test_executor.clear_history()
-        assert len(test_executor.execution_history) == 0
-
-
-class TestTestOrchestrator:
-    """Test TestOrchestrator following TDD principles"""
-
-    @pytest.fixture
-    def test_orchestrator(self):
-        """Create a test orchestrator instance."""
-        return TestOrchestrator()
-
-    @pytest.fixture
-    def sample_suite(self):
-        """Create a sample test suite."""
-        suite = IntegrationTestSuite("Orchestrator Suite", "Orchestrator description")
-        suite.add_test(CrossSystemCommunicationTest("Orchestrator Test", TestPriority.NORMAL))
-        return suite
-
-    @pytest.mark.unit
-    def test_orchestrator_initialization(self, test_orchestrator):
-        """Test orchestrator initialization."""
-        assert test_orchestrator.runner is not None
-        assert test_orchestrator.executor is not None
-        assert test_orchestrator.config["auto_retry_failed"] is True
-
-    @pytest.mark.unit
-    def test_configure_orchestrator(self, test_orchestrator):
-        """Test orchestrator configuration."""
-        config = {"auto_retry_failed": False, "generate_reports": False}
-        test_orchestrator.configure(config)
-        
-        assert test_orchestrator.config["auto_retry_failed"] is False
-        assert test_orchestrator.config["generate_reports"] is False
-
-    @pytest.mark.asyncio
-    async def test_run_test_suite(self, test_orchestrator, sample_suite):
-        """Test running a test suite through orchestrator."""
-        summary = await test_orchestrator.run_test_suite(sample_suite)
-        
-        assert "overall_success_rate" in summary
-        assert summary["total_suites"] >= 0
-        assert summary["total_tests"] >= 0
-
-    @pytest.mark.asyncio
-    async def test_run_multiple_suites(self, test_orchestrator, sample_suite):
-        """Test running multiple test suites through orchestrator."""
-        # Create another suite
-        suite2 = IntegrationTestSuite("Suite 2", "Second suite")
-        suite2.add_test(ServiceIntegrationTest("Service Test", TestPriority.NORMAL))
-        
-        suites = [sample_suite, suite2]
-        summary = await test_orchestrator.run_multiple_suites(suites)
-        
-        assert "overall_success_rate" in summary
-        assert summary["total_suites"] >= 0
-        assert summary["total_tests"] >= 0
-
-
-class TestTestingFrameworkCLI:
-    """Test TestingFrameworkCLI following TDD principles"""
-
-    @pytest.fixture
-    def cli(self):
-        """Create a CLI instance."""
-        return TestingFrameworkCLI()
-
-    @pytest.mark.unit
-    def test_cli_initialization(self, cli):
-        """Test CLI initialization."""
-        assert cli.parser is not None
-        assert cli.orchestrator is not None
-
-    @pytest.mark.unit
-    def test_parser_creation(self, cli):
-        """Test parser creation."""
-        parser = cli._create_parser()
-        assert parser is not None
-
-    @pytest.mark.unit
-    def test_demo_suite_creation(self, cli):
-        """Test demo suite creation."""
-        suite = cli._create_demo_suite_internal("Test Demo Suite")
-        
-        assert suite.name == "Test Demo Suite"
-        assert len(suite.tests) == 3
-        assert suite.description == "Demo test suite: Test Demo Suite"
-
-    @pytest.mark.unit
-    def test_suite_by_name_creation(self, cli):
-        """Test suite creation by name."""
-        # Test communication suite
-        comm_suite = cli._create_suite_by_name("communication")
-        assert comm_suite is not None
-        assert comm_suite.name == "Communication Tests"
-        assert len(comm_suite.tests) == 3
-        
-        # Test service suite
-        service_suite = cli._create_suite_by_name("service")
-        assert service_suite is not None
-        assert service_suite.name == "Service Tests"
-        assert len(service_suite.tests) == 3
-        
-        # Test database suite
-        db_suite = cli._create_suite_by_name("database")
-        assert db_suite is not None
-        assert db_suite.name == "Database Tests"
-        assert len(db_suite.tests) == 3
-        
-        # Test comprehensive suite
-        comp_suite = cli._create_suite_by_name("comprehensive")
-        assert comp_suite is not None
-        assert comp_suite.name == "Comprehensive Tests"
-        assert len(comp_suite.tests) == 3
-
-    @pytest.mark.unit
-    def test_unknown_suite_creation(self, cli):
-        """Test creation of unknown suite."""
-        unknown_suite = cli._create_suite_by_name("unknown")
-        assert unknown_suite is None
-
-    @pytest.mark.unit
-    def test_orchestrator_configuration(self, cli):
-        """Test orchestrator configuration from CLI args."""
-        # Mock args
-        args = Mock()
-        args.retry_failed = True
-        args.parallel = True
-        args.output = "json"
-        
-        cli._configure_orchestrator(args)
-        
-        assert cli.orchestrator.config["auto_retry_failed"] is True
-        assert cli.orchestrator.runner.global_config["parallel_suites"] is True
-        assert cli.orchestrator.runner.global_config["report_format"] == "json"
-
-
-class TestIntegrationScenarios:
-    """Integration tests for complete testing workflows following TDD"""
-
-    @pytest.mark.integration
-    def test_complete_testing_workflow_tdd(self):
-        """Test complete testing workflow from setup to execution following TDD."""
-        # Arrange
-        suite = IntegrationTestSuite("Integration Suite", "Integration test suite")
-        suite.add_test(CrossSystemCommunicationTest("Integration Test", TestPriority.HIGH))
-        
-        # Act
-        # Note: We can't actually run async tests in unit tests without proper setup
-        # This test verifies the structure and configuration
-        
-        # Assert
-        assert suite.name == "Integration Suite"
-        assert len(suite.tests) == 1
-        assert suite.tests[0].test_name == "Integration Test"
-        assert suite.tests[0].priority == TestPriority.HIGH
-
-    @pytest.mark.integration
-    def test_test_suite_management_tdd(self):
-        """Test test suite management workflow following TDD."""
-        # Arrange
-        runner = IntegrationTestRunner()
-        suite1 = IntegrationTestSuite("Suite 1", "First suite")
-        suite2 = IntegrationTestSuite("Suite 2", "Second suite")
-        
-        # Act
-        runner.add_test_suite(suite1)
-        runner.add_test_suite(suite2)
-        
-        # Assert
-        assert len(runner.test_suites) == 2
-        assert runner.test_suites[0].name == "Suite 1"
-        assert runner.test_suites[1].name == "Suite 2"
-        
-        # Test removal
-        runner.remove_test_suite("Suite 1")
-        assert len(runner.test_suites) == 1
-        assert runner.test_suites[0].name == "Suite 2"
-
-    @pytest.mark.integration
-    def test_cli_integration_workflow_tdd(self):
-        """Test CLI integration workflow following TDD."""
-        # Arrange
-        cli = TestingFrameworkCLI()
-        
-        # Act & Assert - Create demo suite
-        suite = cli._create_demo_suite_internal("CLI Test Suite")
-        assert suite.name == "CLI Test Suite"
-        assert len(suite.tests) == 3
-        
-        # Act & Assert - Create suite by name
-        comm_suite = cli._create_suite_by_name("communication")
-        assert comm_suite is not None
-        assert comm_suite.name == "Communication Tests"
-        
-        # Act & Assert - List available suites
-        # This would normally be tested with actual CLI execution
-
-
-if __name__ == "__main__":
-    pytest.main([__file__, "-v"])
